billy_backend/requirements.txt
django>=3.1
djangorestframework
django-rest-knox>=4.1.0